+++ src/TotalRecordsController.php
@@ -43,7 +43,7 @@ class TotalRecordsController extends Controller
                     $seriesSql .= ", SUM(CASE WHEN ".$filter->key." = '".$filter->value."' then ".$calculation." else 0 end) as '".$labelList[$seriesKey]."'";
                 }
             }
-            $query = $model::selectRaw('DATE_FORMAT('.$xAxisColumn.', "%b %y") AS yearmonth, DATE_FORMAT('.$xAxisColumn.', "%y-%m") AS yearmonthorder, count(1) counted'.$seriesSql)
+            $query = $model::selectRaw('DATE_FORMAT('.$xAxisColumn.', "%b %y") AS yearmonth, DATE_FORMAT('.$xAxisColumn.', "%y-%m") AS yearmonthorder, sum('.$calculation.') counted'.$seriesSql)
                 ->where($xAxisColumn, '>=', Carbon::now()->firstOfMonth()->subMonth($dataForLast-1))
                 ->groupBy('yearmonthorder', 'yearmonth')
                 ->orderBy('yearmonthorder', 'asc');